Conflicts:
	palladian-core/src/main/java/ws/palladian/classification/utils/ClassificationUtils.java
	palladian-core/src/test/java/ws/palladian/classification/universal/UniversalClassifierTest.java

+++ palladian-retrieval/src/main/java/ws/palladian/retrieval/search/Searcher.java
@@ -36,7 +36,7 @@ public interface Searcher<R extends SearchResult> {
      * @return
      * @throws SearcherException In case the search fails.
      */
-    int getTotalResultCount(String query) throws SearcherException;
+    long getTotalResultCount(String query) throws SearcherException;
 
     /**
      * <p>
